requirements.txt
aiohttp==3.8.3
aiofiles==22.1.0
python-dotenv==0.21.0